DABeamS.jl

Dynamics and Aeroelasticity of Beam Structures package.

Package Features

Beam finite element model derived from the geometrically-exact theory of anisotropic beams.

Performs steady, trim, dynamic and stability/vibration simulations of beam structures (aircraft, rotors, etc) under aerodynamic loads.

Function Documentation

DABeamS.greeting — Function

greeting()

Greets the user with a message

A more detailed explanation can go here, although I guess it is not needed in this case

Arguments

Notes

Notes can go here

Examples

julia> greeting()
You are using the DABeamS package

DABeamS.plusTwo! — Function

plusTwo!(x::Int64)

Sum the numeric "2" to whatever it receives as input

A more detailed explanation can go here, although I guess it is not needed in this case

Arguments

x: The amount to which we want to add 2

Notes

Notes can go here

Examples

julia> five = plusTwo(3)
5

plusTwo!(x::Float64)

Sum the numeric "2" to whatever it receives as input

A more detailed explanation can go here, although I guess it is not needed in this case

Arguments

x: The amount to which we want to add 2

Notes

Notes can go here

Examples

julia> five = plusTwo(3.0)
5

xcd